crypto_bot/requirements.txt
python-telegram-bot
pandas
pandas-ta
requests
python-dotenv
httpx
openai
scipy
numpy